# twitter-swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    

 
  version: 1.0.11
servers:
  - url: http://localhost:4000
...
paths:
...
  /users/refresh-token:
    post:
      tags:
        - users
      summary: Lấy lại access token
      description: Lấy lại access token
      operationId: refreshToken
      requestBody:
        description: Thông tin lấy lại access token
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwMmNlZmNhNTMzYzQ2Mjc5YTQzMmRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTU2MTYzNTksImV4cCI6MTcwNDI1NjM1OX0.cmFMXDeZAZFl6jQ_ymAqRqW3LB_RZs9oDXTw3C2vx5Q
        required: true
      responses:
        '200':
          description: Lấy lại access token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token success
                  result: 
                    $ref: '#/components/schemas/SuccessAuthentication'
        '422':
          description: Invalid input
...
  /users/resend-verify-email:
    post:
      tags:
        - users
      summary: Gửi lại email xác thực
      description: Gửi lại email xác thực
      operationId: resendVerifyEmail
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Gửi lại email xác thực thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resend verify email success
  /users/forgot_password:
    post:
      tags:
        - users
      summary: Quên mật khẩu
      description: Quên mật khẩu
      operationId: forgotPassword
      requestBody:
        description: Thông tin quên mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: dinhdat.le.2000@gmail.com
        required: true
      responses:
        '200':
          description: Nhận được email reset password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Kiểm tra email để reset mật khẩu
...
  /users/reset_password:
    post:
      tags:
        - users
      summary: Reset mật khẩu
      description: Reset mật khẩu
      operationId: resetPassword
      requestBody:
        description: Thông tin reset mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                forgot_password_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRkY2ZiNTIyZTgzYmFmMjVmYjcyMThjIiwidG9rZW5fdHlwZSI6MiwiaWF0IjoxNjkyNTg4Njk0LCJleHAiOjE2OTMxOTM0OTR9.xjNMRg60gejLoNKb3jbDG1_nFsrP_KDwyo5DXAb9q9c
                password:
                  type: string
                  example: Dat123!
                confirm_password:
                  type: string
                  example: Dat123!
        required: true
      responses:
        '200':
          description: Reset mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset mật khẩu thành công
...
  /users/change-password:
    put:
      tags:
        - users
      summary: Đổi mật khẩu
      description: Đổi mật khẩu
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin đổi mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: Dat123!
                password:
                  type: string
                  example: Dat1234!
                confirm_password:
                  type: string
                  example: Dat1234!
        required: true
      responses:
        default:
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đổi mật khẩu thành công
...
components:
  schemas:
...
    SuccessAuthentication:
        type: object
        properties:
          access_token: 
            type: string
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwMmNlZmNhNTMzYzQ2Mjc5YTQzMmRiIiwidG9rZW5fdHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE2OTU2MTYzNTksImV4cCI6MTY5NTYxNzI1OX0.97QuAjiI0dQLkpK37TlrMO_Atylb944OSa1BWiprdkQ
          refresh_token : 
            type: string
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwMmNlZmNhNTMzYzQ2Mjc5YTQzMmRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTU2MTYzNTksImV4cCI6MTcwNDI1NjM1OX0.cmFMXDeZAZFl6jQ_ymAqRqW3LB_RZs9oDXTw3C2vx5Q
...
  securitySchemes:
    BearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT